+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## Upcoming
 
+- fixed: text sizing with San Francisco
+
+## Upcoming
+
 - added: better image viewing
 - added: iOS 26 tab bar minimize behavior option
 - added: show inline images in modal